Extract timeline item actions into a hook
