+++ src/features/timeline/components/timeline-item/index.tsx
@@ -26,7 +26,6 @@ import {
 } from '@/shared/state/transition-drag';
 import { useMediaLibraryStore } from '@/features/timeline/deps/media-library-store';
 import type { PreviewItemUpdate } from '../../utils/item-edit-preview';
-import { mediaTranscriptionService } from '@/features/timeline/deps/media-transcription-service';
 import { getMediaDragData } from '@/features/timeline/deps/media-library-resolver';
 import { useSettingsStore } from '@/features/timeline/deps/settings';
 import { useTimelineDrag, dragOffsetRef, dragPreviewOffsetByItemRef } from '../../hooks/use-timeline-drag';
@@ -70,14 +69,11 @@ import {
   getTrimOperationBoundsVisual,
 } from './tool-operation-overlay-utils';
 import { useDragVisualState } from './use-drag-visual-state';
+import { useTimelineItemActions } from './use-timeline-item-actions';
 import { AnchorDragGhost, FollowerDragGhost } from './drag-ghosts';
 import { DragBlockedTooltip } from './drag-blocked-tooltip';
 import { ItemContextMenu } from './item-context-menu';
 import { toast } from 'sonner';
-import { useClearKeyframesDialogStore } from '@/shared/state/clear-keyframes-dialog';
-import { useTtsGenerateDialogStore } from '@/shared/state/tts-generate-dialog';
-import type { AnimatableProperty } from '@/types/keyframe';
-import { useBentoLayoutDialogStore } from '../bento-layout-dialog-store';
 import { getRazorSplitPosition } from '../../utils/razor-snap';
 import type { RazorSnapTarget } from '../../utils/razor-snap';
 import { getFilteredItemSnapEdges } from '../../utils/timeline-snap-utils';
@@ -109,7 +105,6 @@ import {
 } from '../../utils/smart-trim-zones';
 import { useMarkersStore } from '../../stores/markers-store';
 import { useCompositionNavigationStore } from '../../stores/composition-navigation-store';
-import { insertFreezeFrame, linkItems, unlinkItems, splitItemAtFrames } from '../../stores/actions/item-actions';
 import {
   createPreComp,
   dissolvePreComp,
@@ -119,19 +114,11 @@ import { useRollHoverStore } from '../../stores/roll-hover-store';
 import { timelineToSourceFrames } from '../../utils/source-calculations';
 import { computeSlideContinuitySourceDelta } from '../../utils/slide-utils';
 import { getTransitionBridgeBounds } from '../../utils/transition-preview-geometry';
-import type { MediaTranscriptModel } from '@/types/storage';
-import { WHISPER_MODEL_LABELS } from '@/shared/utils/whisper-settings';
-import { isLocalInferenceCancellationError } from '@/shared/state/local-inference';
-import { getTranscriptionOverallPercent } from '@/shared/utils/transcription-progress';
 import { getAudioFadePixels, getAudioFadeSecondsFromOffset, type AudioFadeHandle } from '../../utils/audio-fade';
 import { getAudioFadeCurveControlPoint, getAudioFadeCurveFromOffset, getAudioFadeCurvePath } from '../../utils/audio-fade-curve';
 import { getAudioVolumeDbFromDragDelta, getAudioVisualizationScale, getAudioVolumeLineY } from '../../utils/audio-volume';
 import { EDITOR_LAYOUT_CSS_VALUES } from '@/shared/ui/editor-layout';
 import { findHandleNeighborWithTransitions, findNearestNeighbors } from '../../utils/transition-linked-neighbors';
-import { detectScenes } from '../../deps/analysis';
-import { resolveMediaUrl } from '../../deps/media-library-resolver';
-const CAPTION_GENERATION_OVERLAY_ID = 'caption-generation';
-const SCENE_DETECTION_OVERLAY_ID = 'scene-detection';
 const EMPTY_SEGMENT_OVERLAYS = [] as const;
 const ACTIVE_CURSOR_CLASSES = [
   'timeline-cursor-trim-left',
@@ -365,7 +352,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
     handleSlipSlideStart,
   } = useTimelineSlipSlide(item, timelineDuration, trackLocked);
 
-  // Track push functionality — move clip + downstream items to close/open gaps
+  // Track push functionality - move clip + downstream items to close/open gaps
   const { isTrackPushActive, handleTrackPushStart } = useTrackPush(item, timelineDuration, trackLocked);
 
   const activeGlobalCursorClass = useMemo(() => {
@@ -614,7 +601,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
     }, [item.id])
   );
 
-  // Slide range from the preview store — the tightest constraint across all tracks.
+  // Slide range from the preview store - the tightest constraint across all tracks.
   // Used by both primary and companion overlays so limit boxes match.
   const slideRange = useSlideEditPreviewStore(
     useShallow(
@@ -692,7 +679,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
   const right = Math.round(timeToPixels((previewBaseItem.from + previewBaseItem.durationInFrames + slideDurationOffset + slideFromOffset + rippleEditOffset + trackPushOffset) / fps));
   const width = right - left;
 
-  // Source FPS for converting source frames â†' timeline frames (sourceStart etc. are in source-native FPS)
+  // Source FPS for converting source frames -> timeline frames (sourceStart etc. are in source-native FPS)
   const effectiveSourceFps = previewBaseItem.sourceFps ?? fps;
 
   // Preview item for clip internals (filmstrip/waveform) during edit drags.
@@ -819,8 +806,8 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
     let trimVisualLeft = left;
     let trimVisualWidth = width;
 
-    // Ripple edit: compute the new right edge from frames â€” the SAME rounding
-    // path that downstream items use for their `left` â€” so both edges go through
+    // Ripple edit: compute the new right edge from frames - the SAME rounding
+    // path that downstream items use for their `left` - so both edges go through
     // a single Math.round(timeToPixels(totalFrames / fps)) and can never diverge
     // by even 1 px.  `rippleEdgeDelta` equals the downstream `rippleEditOffset`.
     if (rippleEdgeDelta !== 0) {
@@ -1269,7 +1256,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
           const neighbor = findHandleNeighborWithTransitions(item, 'end', items, transitions);
           if (neighbor) useRollHoverStore.getState().setRollHover(item.id, neighbor.id, 'start');
         } else if (prevIntent === 'roll-start' || prevIntent === 'roll-end') {
-          // Was rolling, no longer — clear
+          // Was rolling, no longer - clear
           useRollHoverStore.getState().clearRollHover(item.id);
         }
       }
@@ -1427,8 +1414,8 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
   // Gap detection: clip has empty space before it (no strictly adjacent left neighbor)
   const hasGapBefore = item.from > 0 && !leftNeighbor;
 
-  // Gap width in pixels — used for track-push handle sizing.
-  // Uses getState() snapshot — acceptable because this only affects the push
+  // Gap width in pixels - used for track-push handle sizing.
+  // Uses getState() snapshot - acceptable because this only affects the push
   // handle width, which is only visible during active gap interactions.
   const gapBeforePx = useMemo(() => {
     if (!hasGapBefore) return 0;
@@ -1442,304 +1429,32 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
     return Math.round(frameToPixels(item.from - prevEnd));
   }, [hasGapBefore, item.trackId, item.id, item.from, item.durationInFrames, frameToPixels]);
 
-  // Action handlers
-  const handleJoinSelected = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    if (selectedItemIds.length >= 2) {
-      const itemById = useItemsStore.getState().itemById;
-      const selectedItems = selectedItemIds
-        .map((id) => itemById[id])
-        .filter((i): i is NonNullable<typeof i> => i !== undefined);
-      if (canJoinMultipleItems(selectedItems)) {
-        useTimelineStore.getState().joinItems(selectedItemIds);
-      }
-    }
-  }, []);
-
-  const handleJoinLeft = useCallback(() => {
-    if (leftNeighbor) {
-      useTimelineStore.getState().joinItems([leftNeighbor.id, item.id]);
-    }
-  }, [leftNeighbor, item.id]);
-
-  const handleJoinRight = useCallback(() => {
-    if (rightNeighbor) {
-      useTimelineStore.getState().joinItems([item.id, rightNeighbor.id]);
-    }
-  }, [rightNeighbor, item.id]);
-
-  const handleDelete = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    if (selectedItemIds.length > 0) {
-      useTimelineStore.getState().removeItems(selectedItemIds);
-    }
-  }, []);
-
-  const handleRippleDelete = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    if (selectedItemIds.length > 0) {
-      useTimelineStore.getState().rippleDeleteItems(selectedItemIds);
-    }
-  }, []);
-
-  const handleLinkSelected = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    void linkItems(selectedItemIds);
-  }, []);
-
-  const handleUnlinkSelected = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    unlinkItems(selectedItemIds);
-  }, []);
-
-  const handleClearAllKeyframes = useCallback(() => {
-    useClearKeyframesDialogStore.getState().openClearAll([item.id]);
-  }, [item.id]);
-
-  const handleClearPropertyKeyframes = useCallback((property: AnimatableProperty) => {
-    useClearKeyframesDialogStore.getState().openClearProperty([item.id], property);
-  }, [item.id]);
-
-  // Bento layout
-  const handleBentoLayout = useCallback(() => {
-    const selectedItemIds = useSelectionStore.getState().selectedItemIds;
-    if (selectedItemIds.length < 2) return;
-    useBentoLayoutDialogStore.getState().open(selectedItemIds);
-  }, []);
-
-  // Freeze frame
-  const handleFreezeFrame = useCallback(() => {
-    if (item.type !== 'video') return;
-    const { currentFrame } = usePlaybackStore.getState();
-    void insertFreezeFrame(item.id, currentFrame);
-  }, [item.id, item.type]);
-
-  // Generate audio from text
-  const textContent = item.type === 'text' ? item.text : '';
-  const hasSpeakableText = textContent.trim().length > 0;
-  const handleGenerateAudioFromText = useCallback(() => {
-    if (!hasSpeakableText) return;
-    useTtsGenerateDialogStore.getState().open(textContent, item.id);
-  }, [item.id, textContent, hasSpeakableText]);
-
-  const handleCaptionGeneration = useCallback((
-    model: MediaTranscriptModel,
-    options?: {
-      forceTranscription?: boolean;
-      replaceExisting?: boolean;
-    },
-  ) => {
-    if ((item.type !== 'video' && item.type !== 'audio') || !item.mediaId || isBroken) {
-      return;
-    }
-
-    const mediaId = item.mediaId;
-    const clipId = item.id;
-    const store = useMediaLibraryStore.getState();
-    const overlayStore = useTimelineItemOverlayStore.getState();
-    const previousStatus = store.transcriptStatus.get(mediaId) ?? 'idle';
-    const forceTranscription = options?.forceTranscription ?? false;
-    const replaceExisting = options?.replaceExisting ?? false;
-    const overlayLabel = forceTranscription ? 'Regenerating captions' : 'Generating captions';
-
-    const run = async () => {
-      let updatedTranscriptStatus = previousStatus;
-
-      try {
-        const existingTranscript = await mediaTranscriptionService.getTranscript(mediaId);
-        const needsTranscription =
-          forceTranscription || !existingTranscript || existingTranscript.model !== model;
-
-        if (needsTranscription) {
-          overlayStore.upsertOverlay(clipId, {
-            id: CAPTION_GENERATION_OVERLAY_ID,
-            label: overlayLabel,
-            progress: 0,
-            tone: 'info',
-          });
-          store.setTranscriptStatus(mediaId, 'transcribing');
-          store.setTranscriptProgress(mediaId, { stage: 'loading', progress: 0 });
-
-          await mediaTranscriptionService.transcribeMedia(mediaId, {
-            model,
-            onProgress: (progress) => {
-              const mediaLibraryStore = useMediaLibraryStore.getState();
-              mediaLibraryStore.setTranscriptProgress(mediaId, progress);
-              const mergedProgress = mediaLibraryStore.transcriptProgress.get(mediaId) ?? progress;
-
-              useTimelineItemOverlayStore.getState().upsertOverlay(clipId, {
-                id: CAPTION_GENERATION_OVERLAY_ID,
-                label: overlayLabel,
-                progress: getTranscriptionOverallPercent(mergedProgress),
-                tone: 'info',
-              });
-            },
-          });
-
-          updatedTranscriptStatus = 'ready';
-          store.setTranscriptStatus(mediaId, updatedTranscriptStatus);
-          store.clearTranscriptProgress(mediaId);
-        } else {
-          overlayStore.upsertOverlay(clipId, {
-            id: CAPTION_GENERATION_OVERLAY_ID,
-            label: replaceExisting ? 'Replacing captions' : 'Adding captions',
-            tone: 'info',
-          });
-          updatedTranscriptStatus = 'ready';
-          store.setTranscriptStatus(mediaId, updatedTranscriptStatus);
-          store.clearTranscriptProgress(mediaId);
-        }
-
-        const result = await mediaTranscriptionService.insertTranscriptAsCaptions(mediaId, {
-          clipIds: [clipId],
-          replaceExisting,
-        });
-
-        const successMessage = replaceExisting
-          ? result.insertedItemCount > 0
-            ? result.removedItemCount > 0
-              ? `Replaced ${result.removedItemCount} caption clip${result.removedItemCount === 1 ? '' : 's'} with ${result.insertedItemCount} updated clip${result.insertedItemCount === 1 ? '' : 's'} for this segment using ${WHISPER_MODEL_LABELS[model]}`
-              : `Regenerated ${result.insertedItemCount} caption clip${result.insertedItemCount === 1 ? '' : 's'} for this segment using ${WHISPER_MODEL_LABELS[model]}`
-            : `Removed ${result.removedItemCount} generated caption clip${result.removedItemCount === 1 ? '' : 's'} for this segment using ${WHISPER_MODEL_LABELS[model]}`
-          : `Inserted ${result.insertedItemCount} caption clip${result.insertedItemCount === 1 ? '' : 's'} for this segment with ${WHISPER_MODEL_LABELS[model]}`;
-
-        store.showNotification({
-          type: 'success',
-          message: successMessage,
-        });
-      } catch (error) {
-        if (isLocalInferenceCancellationError(error)) {
-          store.setTranscriptStatus(mediaId, previousStatus);
-          store.clearTranscriptProgress(mediaId);
-          return;
-        }
-
-        store.setTranscriptStatus(mediaId, updatedTranscriptStatus === 'ready' ? 'ready' : 'error');
-        store.clearTranscriptProgress(mediaId);
-        store.showNotification({
-          type: 'error',
-          message: error instanceof Error ? error.message : 'Failed to generate captions for segment',
-        });
-      } finally {
-        useTimelineItemOverlayStore.getState().removeOverlay(clipId, CAPTION_GENERATION_OVERLAY_ID);
-      }
-    };
-
-    void run();
-  }, [item.id, item.mediaId, item.type, isBroken]);
-
-  const handleGenerateCaptions = useCallback((model: MediaTranscriptModel) => {
-    handleCaptionGeneration(model);
-  }, [handleCaptionGeneration]);
-
-  const handleRegenerateCaptions = useCallback((model: MediaTranscriptModel) => {
-    handleCaptionGeneration(model, {
-      forceTranscription: true,
-      replaceExisting: true,
-    });
-  }, [handleCaptionGeneration]);
-
-  const isCaptionGenerationActive = segmentOverlays.some(
-    (overlay) => overlay.id === CAPTION_GENERATION_OVERLAY_ID,
-  );
-
-  const isSceneDetectionActive = segmentOverlays.some(
-    (overlay) => overlay.id === SCENE_DETECTION_OVERLAY_ID,
-  );
-
-  const handleDetectScenes = useCallback(() => {
-    if (item.type !== 'video' || !item.mediaId || isBroken) return;
-
-    const mediaId = item.mediaId;
-    const clipId = item.id;
-    const clipFrom = item.from;
-    const overlayStore = useTimelineItemOverlayStore.getState();
-
-    const run = async () => {
-      const abortController = new AbortController();
-
-      try {
-        overlayStore.upsertOverlay(clipId, {
-          id: SCENE_DETECTION_OVERLAY_ID,
-          label: 'Detecting scenes',
-          progress: 0,
-          tone: 'info',
-        });
-
-        const url = await resolveMediaUrl(mediaId);
-        const video = document.createElement('video');
-        video.src = url;
-        video.muted = true;
-        video.preload = 'auto';
-
-        await new Promise<void>((resolve, reject) => {
-          video.onloadedmetadata = () => resolve();
-          video.onerror = () => reject(new Error('Failed to load video for scene detection'));
-        });
-
-        const currentFps = useTimelineStore.getState().fps;
-        const cuts = await detectScenes(video, currentFps, {
-          sampleIntervalMs: 500,
-          useGemmaVerification: false,
-          signal: abortController.signal,
-          onProgress: (progress) => {
-            const stageLabels = {
-              'optical-flow': `Analyzing motion (${progress.sceneCuts} candidates)`,
-              'loading-model': `Loading Gemma model (${progress.percent}%)`,
-              'verifying': `Verifying cuts (${progress.sceneCuts}/${progress.totalSamples} confirmed)`,
-            };
-            const label = stageLabels[progress.stage ?? 'optical-flow'];
-            useTimelineItemOverlayStore.getState().upsertOverlay(clipId, {
-              id: SCENE_DETECTION_OVERLAY_ID,
-              label,
-              progress: progress.stage === 'loading-model' ? progress.percent : progress.percent,
-              tone: 'info',
-            });
-          },
-        });
-
-        video.src = '';
-
-        if (cuts.length === 0) {
-          toast.info('No scene cuts detected');
-          return;
-        }
-
-        // Convert scene cut frames to timeline split frames (relative to clip start).
-        // Scene detection returns frames in source video space — offset by clip's `from`.
-        // Filter to cuts within the clip's duration range.
-        const clipDuration = item.durationInFrames;
-        const sourceStartFrame = item.sourceStart ?? 0;
-        const splitFrames = cuts
-          .map((cut) => cut.frame - sourceStartFrame)
-          .filter((f) => f > 0 && f < clipDuration)
-          .map((f) => f + clipFrom);
-
-        if (splitFrames.length === 0) {
-          toast.info('No scene cuts within clip bounds');
-          return;
-        }
-
-        const splitCount = splitItemAtFrames(clipId, splitFrames);
-
-        if (splitCount > 0) {
-          toast.success(`Split clip at ${splitCount} scene cut${splitCount > 1 ? 's' : ''}`);
-        } else {
-          toast.info('No valid split points found');
-        }
-      } catch (err) {
-        if (err instanceof Error && err.message.includes('WebGPU')) {
-          toast.error('Scene detection requires WebGPU support');
-        } else {
-          toast.error('Scene detection failed');
-        }
-      } finally {
-        useTimelineItemOverlayStore.getState().removeOverlay(clipId, SCENE_DETECTION_OVERLAY_ID);
-      }
-    };
-
-    void run();
-  }, [item.id, item.type, item.mediaId, item.from, item.durationInFrames, (item as { sourceStart?: number }).sourceStart, isBroken]);
+  const {
+    hasSpeakableText,
+    isCaptionGenerationActive,
+    isSceneDetectionActive,
+    handleJoinSelected,
+    handleJoinLeft,
+    handleJoinRight,
+    handleDelete,
+    handleRippleDelete,
+    handleLinkSelected,
+    handleUnlinkSelected,
+    handleClearAllKeyframes,
+    handleClearPropertyKeyframes,
+    handleBentoLayout,
+    handleFreezeFrame,
+    handleGenerateAudioFromText,
+    handleGenerateCaptions,
+    handleRegenerateCaptions,
+    handleDetectScenes,
+  } = useTimelineItemActions({
+    item,
+    isBroken,
+    leftNeighbor,
+    rightNeighbor,
+    segmentOverlays,
+  });
 
   // Composition operations
   const isCompositionItem = item.type === 'composition' || (item.type === 'audio' && !!item.compositionId);
@@ -2393,7 +2108,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
       const committedVolume = audioVolumePreviewRef.current ?? latestPreviewVolume;
       audioVolumeCleanupRef.current?.();
       audioVolumeCleanupRef.current = null;
-      // Keep live gain active — segment volumeDb is stale until composition
+      // Keep live gain active - segment volumeDb is stale until composition
       // naturally re-renders, and the audio component auto-clears via useEffect.
 
       finalizeAudioVolumeChange(committedVolume, {
@@ -2582,7 +2297,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
   ), [linkedItemsForSync, linkedSyncPreviewItem, fps, linkedSyncPreviewUpdatesById, suppressLinkedSyncBadge]);
 
   const handleCreatePreComp = useCallback(() => {
-    // Capture selection synchronously â€” context menu close may clear it before the dynamic import resolves
+    // Capture selection synchronously - context menu close may clear it before the dynamic import resolves
     const ids = useSelectionStore.getState().selectedItemIds;
     createPreComp(undefined, ids);
   }, []);
@@ -2996,7 +2711,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
                 ? getAudioVisualizationScale(audioVolumePreviewRef.current)
                 : audioVisualizationScale
             ),
-          }}
+          } as React.CSSProperties}
           onClick={handleClick}
           onDoubleClick={handleDoubleClick}
           onMouseDown={handleMouseDown}
@@ -3008,7 +2723,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
           onDragLeave={handleEffectDragLeave}
           onDrop={handleEffectDrop}
         >
-          {/* Selection indicator — hidden during active gestures to reduce clutter */}
+          {/* Selection indicator - hidden during active gestures to reduce clutter */}
           {isSelected && !trackLocked && !isAnyGestureActive && (
             <div className="absolute inset-0 rounded pointer-events-none z-20 border border-primary" />
           )}
@@ -3252,7 +2967,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
         </div>
       </ItemContextMenu>
 
-      {/* Track push handle — sits in the gap to the LEFT of the clip, outside contain:paint */}
+      {/* Track push handle - sits in the gap to the LEFT of the clip, outside contain:paint */}
       <TrackPushHandle
         enabled={hasGapBefore && !trackLocked && activeTool === 'trim-edit'}
         isActive={isTrackPushActive}
@@ -3263,7 +2978,7 @@ export const TimelineItem = memo(function TimelineItem({ item, timelineDuration 
 
       <ToolOperationOverlay visual={toolOperationOverlay} />
 
-      {/* Active edge halos — top layer, above both clip and bounds box */}
+      {/* Active edge halos - top layer, above both clip and bounds box */}
       {activeEdges && (
         <div
           className="absolute inset-y-0 pointer-events-none"
